requirements.txt
python-telegram-bot==20.3
qrcode
pyotp
python-dotenv
psycopg2-binary
